Backend/requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
click==8.1.3
Django==4.2.1
graphql-core==3.2.3
h11==0.14.0
idna==3.4
libcst==0.4.9
markdown-it-py==2.2.0
mdurl==0.1.2
mypy-extensions==1.0.0
Pygments==2.15.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0
redis==4.5.4
rich==13.3.5
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.26.1
strawberry-graphql==0.176.3
strawberry-graphql-django==0.9.4
typing-inspect==0.8.0
typing_extensions==4.5.0
uvicorn==0.21.1
